Updates 3DS2 Fingerprinting action to include directory certificate (#1207)
* Updates 3DS2 Fingerprinting action to include directory certificate.

* Updates SDK.

* Without the logs.

* Update param name.

* Updates 3DS2 SDK

* Update with latest 3DS2 framework build.

* Apply suggestions from code review

Use stp_stringForKey


* Update Stripe/STPIntentActionUseStripeSDK.m

Fix typo

+++ Stripe/Payments/STPThreeDSTextFieldCustomization.m
@@ -25,11 +25,11 @@
     return self;
 }
 
-- (NSInteger)borderWidth {
+- (CGFloat)borderWidth {
     return self.textFieldCustomization.borderWidth;
 }
 
-- (void)setBorderWidth:(NSInteger)borderWidth {
+- (void)setBorderWidth:(CGFloat)borderWidth {
     self.textFieldCustomization.borderWidth = borderWidth;
 }
